+++ front/components/actions/request_user_data/RequestUserDataActionDetails.tsx
